Rag+Gradio/requirements.txt
gradio==4.7.1
openai==2.26.0
python-dotenv==1.0.0
loguru==0.7.2
qdrant_client==1.16.1
pdfplumber==0.10.3
langchain>=0.3.0
tiktoken==0.5.2
pandas==2.3.3
openpyxl==3.1.2